refactor: Update UserInput component to use controlled inputs
