+++ devhub/projects/udemy_react_the_complete_guide_2024/s05/src/App.jsx
@@ -1,11 +1,31 @@
+import { useState } from "react";
+
 import Header from "./components/Header";
 import UserInput from "./components/UserInput";
+import Results from "./components/Results";
 
 function App() {
+  const [userInput, setUserInput] = useState({
+    initialInvestment: 10000,
+    annualInvestment: 1200,
+    expectedReturn: 6,
+    duration: 10,
+  });
+
+  function handleChange(inputIdentifier, newValue) {
+    setUserInput((prevState) => {
+      return {
+        ...prevState,
+        [inputIdentifier]: newValue,
+      };
+    });
+  }
+
   return (
     <>
       <Header />;
-      <UserInput />;
+      <UserInput userInput={userInput} onChange={handleChange} />;
+      <Results input={userInput} />;
     </>
   );
 }
