+++ devhub/projects/udemy_react_the_complete_guide_2024/s05/src/components/Results.jsx
@@ -1,0 +1,4 @@
+export default function Results({ input }) {
+  console.log(input);
+  return <p>Results...</p>;
+}
